moved message processing to FromDiscordExtensions
