+++ StackoverflowChatbot/Discord.cs
@@ -1,14 +1,12 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using Discord;
 using Discord.WebSocket;
 using SharpExchange.Chat.Actions;
 using SharpExchange.Chat.Events;
 using SharpExchange.Net.WebSockets;
+using StackoverflowChatbot.Relay;
 
 namespace StackoverflowChatbot
 {
@@ -58,7 +56,7 @@ namespace StackoverflowChatbot
 					var displayname = string.IsNullOrEmpty(user.Nickname) ? user.Username : user.Nickname;
 
 					// Create a list of messages in case there are embedded codeblocks or stuff alongside text
-					var messages = BuildSoMessage(user, config, arg);
+					var messages = FromDiscordExtensions.BuildSoMessage(user, config, arg);
 					
 					foreach (var message in messages)
 					{
@@ -93,49 +91,6 @@ namespace StackoverflowChatbot
 			}
 
 			return Task.CompletedTask;
-		}
-
-        private static List<string> BuildSoMessage(SocketGuildUser user, Config.Base config, SocketMessage arg)
-        {
-			var displayname = string.IsNullOrEmpty(user.Nickname) ? user.Username : user.Nickname;
-            var messageStart = $@"\[**[{displayname}]({config.DiscordInviteLink})**] ";
-			var messageContent = arg.Content;
-			var result = new List<string>();
-
-			foreach (var mentionedUser in arg.MentionedUsers)
-			{
-				messageContent = messageContent.Replace(mentionedUser.Mention, $"@{mentionedUser.Username}");
-			}
-			foreach (var mentionedRoles in arg.MentionedRoles)
-			{
-				messageContent = messageContent.Replace(mentionedRoles.Mention, $"[@{mentionedRoles.Name}]({config.DiscordInviteLink})");
-			}
-			foreach (var mentionedChannel in arg.MentionedChannels)
-			{
-				// Library doesn't provide channel mention string
-				messageContent = messageContent.Replace($"<#{mentionedChannel.Id}>", $"[#{mentionedChannel.Name}]({config.DiscordInviteLink})");
-			}
-			
-			var embeddedCode = Regex.Matches(messageContent, "```.+?```", RegexOptions.Singleline);
-			if (embeddedCode.Count == 0)
-				return new List<string>() { messageStart + messageContent };
-
-			// Complex message with codeblocks...
-			int cursor = 0;
-			result.Add(messageStart);
-			foreach (Match codeBlock in embeddedCode)
-			{
-				var indicatorCount = Regex.Matches(messageContent.Substring(0, codeBlock.Index + 3), "```", RegexOptions.Singleline).Count;
-				if(indicatorCount % 2 == 0)
-					continue;
-
-				var soCodeBlock = "    " + codeBlock.ToString().Replace("`", "").Replace("\n", "\n    ").TrimEnd();
-				result.Add(messageContent.Substring(cursor, codeBlock.Index - cursor));
-				cursor = codeBlock.Index + codeBlock.Length;
-				result.Add(soCodeBlock);
-			}
-
-			return result;
-        }
+		}       
     }
 }
